Move insights into a separate file.
