+++ src/main/resources/META-INF/plugin.xml
@@ -100,16 +100,16 @@
         <!-- Remote repo input -->
         <fileType
                 name="SourcegraphRemoteRepoListFile"
-                implementationClass="com.sourcegraph.cody.context.ui.RemoteRepoFileType"
+                implementationClass="com.sourcegraph.cody.context.RemoteRepoFileType"
                 fieldName="INSTANCE"
                 language="SourcegraphRemoteRepoList"
-                extensions="todoremove"/>
+                extensions=""/>
         <lang.parserDefinition
                 language="SourcegraphRemoteRepoList"
-                implementationClass="com.sourcegraph.cody.context.ui.RemoteRepoListParserDefinition"/>
-        <completion.contributor language="SourcegraphRemoteRepoList" implementationClass="com.sourcegraph.cody.context.ui.RemoteRepoCompletionContributor"
+                implementationClass="com.sourcegraph.cody.context.RemoteRepoListParserDefinition"/>
+        <completion.contributor language="SourcegraphRemoteRepoList" implementationClass="com.sourcegraph.cody.context.RemoteRepoCompletionContributor"
                                 order="first, before commitCompletion"/>
-        <annotator language="SourcegraphRemoteRepoList" implementationClass="com.sourcegraph.cody.context.ui.RemoteRepoAnnotator" />
+        <annotator language="SourcegraphRemoteRepoList" implementationClass="com.sourcegraph.cody.context.RemoteRepoAnnotator" />
     </extensions>
 
     <applicationListeners>
